Rename package finntech to finn
