+++ pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>no.finntech.retriable-kafka-consumer</groupId>
+    <groupId>no.finn.retriable-kafka-consumer</groupId>
     <artifactId>retriable-kafka-consumer</artifactId>
     <packaging>jar</packaging>
     <version>1.50-SNAPSHOT</version>
