+++ src/main/java/no/finn/retriableconsumer/ReliableKafkaConsumerPool.java
@@ -1,4 +1,4 @@
-package no.finntech.retriableconsumer;
+package no.finn.retriableconsumer;
 
 import java.io.Closeable;
 import java.util.List;
