+++ src/main/java/no/finn/retriableconsumer/ReliablePoolBuilder.java
@@ -1,11 +1,11 @@
-package no.finntech.retriableconsumer;
+package no.finn.retriableconsumer;
 
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.function.Function;
 
-import no.finntech.retriableconsumer.version.ExposeVersion;
+import no.finn.retriableconsumer.version.ExposeVersion;
 
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
