+++ src/main/java/no/finn/retriableconsumer/Restartable.java
@@ -1,4 +1,4 @@
-package no.finntech.retriableconsumer;
+package no.finn.retriableconsumer;
 
 import java.io.Closeable;
