+++ src/test/java/no/finn/retriableconsumer/RestartableKafkaConsumerTest.java
@@ -1,4 +1,4 @@
-package no.finntech.retriableconsumer;
+package no.finn.retriableconsumer;
 
 import java.util.Arrays;
 import java.util.HashMap;
