+++ src/test/java/no/finn/retriableconsumer/TestUtilTest.java
@@ -1,4 +1,4 @@
-package no.finntech.retriableconsumer;
+package no.finn.retriableconsumer;
 
 import org.junit.Test;
