+++ src/test/kotlin/no/finn/retriableconsumer/ReliableKafkaConsumerPoolTest.kt
@@ -1,4 +1,4 @@
-package no.finntech.retriableconsumer
+package no.finn.retriableconsumer
 
 import org.apache.kafka.clients.consumer.Consumer
 import org.apache.kafka.clients.consumer.ConsumerRecord
